requirements.txt
asgiref==3.5.2
Django==4.1.3
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.6
